shape kind of file, move it to write package
