--- .vscode/extensions.json
@@ -1,5 +1,0 @@
-{
-    "recommendations": [
-        "ttoowa.in-your-face-incredible"
-    ]
-}
