+++ domain/item/item.go
@@ -1,6 +1,10 @@
 package item
 
-import "github.com/cezarovici/goLayouter/domain"
+import (
+	"strings"
+
+	"github.com/cezarovici/goLayouter/domain"
+)
 
 type KindOfFile int8
 
@@ -23,3 +27,28 @@ type Item struct {
 
 // Items represents a collection of items.
 type Items []Item
+
+// KindOfFile returns a string representing the kind of a file based on its name.
+func NewKindOfFile(fileName string) KindOfFile {
+	if !strings.Contains(fileName, ".") {
+		return Folder
+	}
+
+	if fileName == "main" || fileName == "main.go" {
+		return Main
+	}
+
+	if strings.Contains(fileName, "test") {
+		return Test
+	}
+
+	if strings.Contains(fileName, "obj") {
+		return Object
+	}
+
+	if strings.Contains(fileName, ".") {
+		return NormalFile
+	}
+
+	return Folder
+}
